requirements.txt
fastapi
uvicorn

langchain
langchain-community

sentence-transformers
faiss-cpu

streamlit
requests

python-multipart
python-dotenv
datasets
pandas
numpy

chromadb==0.5.5
pydantic==2.7.4
pydantic-settings==2.3.4
